python-game/requirements.txt
pygame==2.1.2
pyserial
pytz
pandas
numpy
scipy
plotly
kaleido
fpdf